split some APIs from nn.py to sequence_lod.py
test=develop
