+++ python/paddle/fluid/layers/__init__.py
@@ -35,6 +35,7 @@ from .metric_op import *
 from .learning_rate_scheduler import *
 from .collective import *
 from .distributions import *
+from .sequence_lod import *
 from . import rnn
 
 __all__ = []
@@ -48,6 +49,7 @@ __all__ += detection.__all__
 __all__ += metric_op.__all__
 __all__ += learning_rate_scheduler.__all__
 __all__ += distributions.__all__
+__all__ += sequence_lod.__all__
 __all__ += rnn.__all__
 
 from .rnn import *
